Исправил удаление групп товаров

+++ controllers/GroupsController.php
@@ -113,12 +113,17 @@ class GroupsController extends Controller
         foreach ($goods as $good) {
 
             $path = Yii::getAlias("@app/web/uploads/images/" . $good['id']);
-            $files = array_diff(scandir($path), array('.','..'));
 
-            foreach ($files as $file) {
-                (is_dir("$path/$file")) ? rmdir("$path/$file") : unlink("$path/$file");
+            if (file_exists($path)) {
+
+                $files = array_diff(scandir($path), array('.', '..'));
+
+                foreach ($files as $file) {
+                    (is_dir("$path/$file")) ? rmdir("$path/$file") : unlink("$path/$file");
+                }
+
+                rmdir($path);
             }
-            rmdir($path);
         }
 
         $this->findModel($id)->delete();
